dissolve clipper into clipper+
