+++ bindings/matlab/clique_optimization_mex.cpp
@@ -42,12 +42,11 @@ const int nnodes = (useSparse) ? adjs.rows() : adj.rows();
 
 const auto t1 = std::chrono::high_resolution_clock::now(); // timer
 
-// run clique_optimization:
-long unsigned int clique_size = 0;
-std::vector<long int> clique;
+
 Eigen::MatrixXd M = adj + Eigen::MatrixXd::Identity(nnodes,nnodes);
 Eigen::VectorXd u0 = randvec(nnodes);
-clipperplus::clique_optimization(M, u0, clique_size, clique);
+std::vector<long int> clique = clipperplus::clique_optimization(M, u0, clipperplus::Params());
+long long int clique_size = clique.size();
 
 const auto t2 = std::chrono::high_resolution_clock::now();
 const auto duration = std::chrono::duration_cast<std::chrono::microseconds>(t2 - t1);
