--- include/clipper/invariants/builtins.h
@@ -1,11 +1,0 @@
-/**
- * @file builtins.h
- * @brief Common geometric invariants for CLIPPER
- * @author Parker Lusk <plusk@mit.edu>
- * @date 3 October 2020
- */
-
-#pragma once
-
-#include "clipper/invariants/euclidean_distance.h"
-#include "clipper/invariants/pointnormal_distance.h"
