+++ src/utils.cpp
@@ -1,35 +1,37 @@
 
 #include "clipperplus/utils.h"
 
-namespace clipperplus {
+namespace clipperplus::utils
+{
 
-/** find the index of an element in an std::vector of integers 
-* that is equal to a given value by iterating through the 
-* vector and checking each element. */
-int find_index(const std::vector<int>& vec, int val) {
-    for (int i = 0; i < vec.size(); ++i) {
-        if (vec[i] == val) {
-            return i; // Return the index of the first occurrence of 'val'
+Eigen::VectorXd selectFromIndicator(
+    const Eigen::VectorXd &x,
+    const Eigen::VectorXi &ind)
+{
+    Eigen::VectorXd y(ind.sum());
+    size_t idx = 0;
+    for (size_t i = 0; i < x.size(); ++i)
+    {
+        if (ind[i])
+        {
+            y[idx] = x[i];
+            idx++;
         }
     }
-    return -1; // Return -1 if 'val' is not found in the vector
+    return y;
 }
 
-// convert adjacency matrix to adjacency list
-void adjmat_to_adjlist(const Eigen::MatrixXd& adj,
-                       const int& nnodes,
-                       int* ei,
-                       int* ej) {
-    int counter = 0;
-    for (int i=0; i<nnodes-1; i++) {
-        for (int j=i+1; j<nnodes; j++) {
-        if (adj(i,j)==1) {
-                ei[counter] = i;
-                ej[counter] = j;
-                counter++;
-            }
-        }
-    }
-} 
+std::vector<long> findIndicesWhereAboveThreshold(
+    const Eigen::VectorXd& x,
+    double thr
+) {
+  std::vector<long> indices;
+  indices.reserve(x.size());
+  for (size_t i=0; i<x.rows(); ++i) {
+    if (x(i) > thr) indices.push_back(i);
+  }
+  return indices;
+}
 
-} 
+
+}
